+++ {{ cookiecutter.project_name }}/composition_repository/django_filer/settings.py
@@ -19,5 +19,4 @@ class DjangoFilerSettings(EnabledApplicationMarker):
         cls.INSTALLED_APPS.extend([
             "easy_thumbnails",
             "filer",
-            "mptt",
         ])