+++ docs/history.rst
@@ -11,6 +11,8 @@ Version 0.3.8 - Unreleased
 * Improved how elligible Django application modules are discovered in Makefile tasks
   that need it. This should fix issue with some system that don't have a complete
   support of all ``ls`` arguments so it has been written in a full Python script;
+* Upgraded to ``django-filer>=3`` and remove its dependancy to ``mptt`` that are no
+  longer needed;
 
 
 Version 0.3.7 - 2023/06/06